+++ src/sandbox/enviroments/twenty_forty_eight/game/state.py
@@ -1,4 +1,4 @@
-from sandbox.enviroments.twenty_forty_eight.base.state import State
+from sandbox.enviroments.twenty_forty_eight.game.base.state import State
 from dataclasses import dataclass
 from typing import Tuple, cast
 from numpy.typing import NDArray
